Fix for synced files in Vagrant
- Changed box to use centos 7. May use another box in the future
but this will do for now. Something is going on with the how Windows
syncs and mounts files in Virtualbox. I've tried several different
boxes, various versions of VirtualBox and Guest Additions, I tried to
manually install it via the provision script to download, mount and install the
guest additions iso and much more. None of that worked. I speat 2 weeks
trying to solve this issue only to run into more issues with what I
thought were solutions.
- Reduced the vb memory to 4 GB instead of 8 GB.
- Added permission to execute provision script.
- Updated the README file to reflect new process.